Optimizations

+++ framework/aregextend/console/private/SystemServiceConsole.cpp
@@ -109,13 +109,13 @@ inline void SystemServiceConsole::_output_data_rate()
     {
         DataRateHelper::DataRate rateSend{ mDataRateHelper->query_bytes_sent_with_literals() };
         DataRateHelper::DataRate rateRecv{ mDataRateHelper->query_bytes_received_with_literals() };
-        const uint64_t msgsSent = mDataRateHelper->query_msgs_sent();
-        const uint64_t msgsRecv = mDataRateHelper->query_msgs_received();
+        const uint32_t msgsSent = mDataRateHelper->query_msgs_sent();
+        const uint32_t msgsRecv = mDataRateHelper->query_msgs_received();
 
         console.output_msg( areg::ext::COORD_SEND_RATE, areg::ext::FORMAT_SEND_DATA.data( ), static_cast<double>(rateSend.first), rateSend.second.c_str( ) );
         console.output_msg( areg::ext::COORD_RECV_RATE, areg::ext::FORMAT_RECV_DATA.data( ), static_cast<double>(rateRecv.first), rateRecv.second.c_str( ) );
-        console.output_msg( areg::ext::COORD_SEND_MSGS, areg::ext::FORMAT_SEND_MSGS.data( ), static_cast<uint32_t>(msgsSent) );
-        console.output_msg( areg::ext::COORD_RECV_MSGS, areg::ext::FORMAT_RECV_MSGS.data( ), static_cast<uint32_t>(msgsRecv) );
+        console.output_msg( areg::ext::COORD_SEND_MSGS, areg::ext::FORMAT_SEND_MSGS.data( ), msgsSent );
+        console.output_msg( areg::ext::COORD_RECV_MSGS, areg::ext::FORMAT_RECV_MSGS.data( ), msgsRecv );
         console.refresh_screen( );
     }
