+++ framework/logcollector/app/private/LogCollector.cpp
@@ -216,16 +216,12 @@ void LogCollector::run_console_input_simple()
 
             DataRateHelper::DataRate sent = helper.query_bytes_sent_with_literals();
             DataRateHelper::DataRate recv = helper.query_bytes_received_with_literals();
-            const uint64_t msgs_sent = helper.query_msgs_sent();
-            const uint64_t msgs_recv = helper.query_msgs_received();
-            console.output_msg(areg::ext::COORD_SEND_RATE, areg::ext::FORMAT_SEND_DATA.data(),
-                               static_cast<double>(sent.first), sent.second.c_str());
-            console.output_msg(areg::ext::COORD_RECV_RATE, areg::ext::FORMAT_RECV_DATA.data(),
-                               static_cast<double>(recv.first), recv.second.c_str());
-            console.output_msg(areg::ext::COORD_SEND_MSGS, areg::ext::FORMAT_SEND_MSGS.data(),
-                               static_cast<uint32_t>(msgs_sent));
-            console.output_msg(areg::ext::COORD_RECV_MSGS, areg::ext::FORMAT_RECV_MSGS.data(),
-                               static_cast<uint32_t>(msgs_recv));
+            const uint32_t msgs_sent = helper.query_msgs_sent();
+            const uint32_t msgs_recv = helper.query_msgs_received();
+            console.output_msg(areg::ext::COORD_SEND_RATE, areg::ext::FORMAT_SEND_DATA.data(), static_cast<double>(sent.first), sent.second.c_str());
+            console.output_msg(areg::ext::COORD_RECV_RATE, areg::ext::FORMAT_RECV_DATA.data(), static_cast<double>(recv.first), recv.second.c_str());
+            console.output_msg(areg::ext::COORD_SEND_MSGS, areg::ext::FORMAT_SEND_MSGS.data(), msgs_sent);
+            console.output_msg(areg::ext::COORD_RECV_MSGS, areg::ext::FORMAT_RECV_MSGS.data(), msgs_recv);
             console.refresh_screen();
         }
     });
